total 값 long 값 변환 로그 남기기

+++ src/main/java/org/l3cache/dao/PostManager.java
@@ -107,7 +107,7 @@ public class PostManager {
 		return sqlSession.selectOne("PostMapper.countUserPostsList", uid);
 	}
 
-	public int getUserLikesCount(int uid) {
+	public long getUserLikesCount(int uid) {
 		return sqlSession.selectOne("PostMapper.countUserLikesList", uid);
 	}
